requirements.txt
matplotlib
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
attrs==25.3.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
cryptography==45.0.3
defusedxml==0.7.1
discord==2.3.2
discord.py==2.5.2
distro==1.9.0
docstring_parser==0.16
elevenlabs==2.1.0
et_xmlfile==2.0.0
flatbuffers==25.2.10
frozenlist==1.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
groq==0.26.0
grpcio==1.72.1
grpcio-status==1.71.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
instructor==1.8.3
Jinja2==3.1.6
jiter==0.8.2
magika==0.6.2
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.4
numpy==2.2.6
onnxruntime==1.22.0
openai==1.83.0
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pdfminer.six==20250506
pillow==11.2.1
portalocker==2.10.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
PyNaCl==1.5.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
qdrant-client==1.14.2
redis==6.2.0
requests==2.32.3
rich==13.9.4
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sympy==1.14.0
tenacity==9.1.2
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
websockets==15.0.1
yarl==1.20.0
youtube-transcript-api==1.0.3
